\name{Internal Functions}
\alias{.binary.activematrix.check}
\alias{.build.button.function}
\alias{.build.command.argument}
\alias{.eval.command}
\alias{.find.frame}
\alias{.give.doublequote}
\alias{.is.binary.matrix}
\alias{.make.correct.frame}
\alias{.order.button.frames}
\alias{.output.sparse.txt}
\alias{.rcmdr.warning}
\alias{.transform.vector2text}
\alias{.update.biclustering.object}
\alias{Setwd}


\title{List of Internal Functions}

\description{List of internal functions used in various functions in the package.}
\usage{
.binary.activematrix.check()
.build.button.function(function.command,arg.names,button_result,new.frames,save)
.build.command.argument(current.frame,command)
.eval.command(x)
.find.frame(x,frame.name)
.give.doublequote(x)
.is.binary.matrix(x)
.make.correct.frame(title,border,window)
.order.button.frames(new.frames)
.output.sparse.txt(X,file)
.rcmdr.warning(x)
Setwd(x=TRUE)
.transform.vector2text(x)
.update.biclustering.object(object,where="all")

}
%\arguments{
%\item{function.command}{ }
%\item{arg.names}{ }
%\item{button_result}{ }
%\item{new.frames}{ }
%\item{save}{ }
%\item{current.frame}{ }
%\item{command}{ }
%\item{x}{ }
%\item{frame.name}{ }
%\item{title}{ }
%\item{border}{ }
%\item{window}{ }
%\item{X}{ }
%\item{file}{ }
%\item{object}{ }
%\item{where}{ }
%}

\details{Not to be used by the users.}


\author{Ewoud De Troyer \email{ewoud.detroyer@student.uhasselt.be}}

