\name{BiclustGUI-internal}
\title{Internal BiclustGUI objects}
\alias{parallelCoordinates2}
\alias{PlaidWin}
\alias{BCCWin}
\alias{BimaxWin}
\alias{PlotWin}
\alias{XmotifsWin}
\alias{SpectralWin}
\alias{QuestWin}
\description{Internal BiclustGUI objects.}
\details{These are not to be called by the user.}
\keyword{internal}

